Renames aws package to database Splits database module into AWS and non-AWS packages Fixes files ot support refactor
