--- packages/aws/package.go
@@ -1,1 +1,0 @@
-package aws
