+++ packages/database/repository.go
@@ -1,0 +1,6 @@
+package database
+
+type Repository interface {
+	Get(id string, dest interface{}) error
+	// Set(id string, source interface{}) error
+}
