+++ packages/tokens/repositories/dynamo_token_repository_test.go
@@ -9,13 +9,13 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	t_dynamo "theodo.red/creditcompanion/packages/aws/dynamodb"
+	"theodo.red/creditcompanion/packages/database/tdynamo"
 	"theodo.red/creditcompanion/packages/tokens/models"
 )
 
 func TestGetsAToken(t *testing.T) {
 	// Given a mock dynamodb
-	dynamoDBMock := new(t_dynamo.DynamoDbMock)
+	dynamoDBMock := new(tdynamo.DynamoDbMock)
 
 	// And given a mock db response
 
@@ -64,7 +64,7 @@ func TestGetsAToken(t *testing.T) {
 
 func TestAnnotatesDbRequestCallError(t *testing.T) {
 	// Given the db returns an error when getting the token
-	dynamoDBMock := new(t_dynamo.DynamoDbMock)
+	dynamoDBMock := new(tdynamo.DynamoDbMock)
 
 	mockError := errors.New("Mock error.")
 	dynamoDBMock.On("GetItem", mock.Anything).Return(nil, mockError)
@@ -83,7 +83,7 @@ func TestAnnotatesDbRequestCallError(t *testing.T) {
 
 func TestReturnsAnErrorIfTheItemCannotBeFound(t *testing.T) {
 	// Given the db returns a nil item
-	dynamoDBMock := new(t_dynamo.DynamoDbMock)
+	dynamoDBMock := new(tdynamo.DynamoDbMock)
 
 	dynamoDBMock.On("GetItem", mock.Anything).Return(&dynamodb.GetItemOutput{Item: nil}, nil)
 
@@ -101,7 +101,7 @@ func TestReturnsAnErrorIfTheItemCannotBeFound(t *testing.T) {
 
 func TestReturnsAnErrorIfUnmarshallingReturnsNullValueToken(t *testing.T) {
 	// Given the db returns unexpected data
-	dynamoDBMock := new(t_dynamo.DynamoDbMock)
+	dynamoDBMock := new(tdynamo.DynamoDbMock)
 
 	mockResultValue := "Mock result value"
 	mockResponse := &dynamodb.GetItemOutput{
